# docs/api.yaml
openapi: 3.0.3
info:
  title: AidBridge API
  description: |
    AidBridge - Disaster Relief Coordinator API

    ## Overview
    RESTful API for coordinating disaster relief operations including:
    - SOS rescue requests
    - Aid supply requests and delivery
    - Volunteer mission management
    - Hub inventory management
    - Sponsor donations

    ## Authentication
    All protected endpoints require Bearer token in Authorization header.

    ## Database Version
    This API is aligned with database schema v4.0

    ## Caching Strategy

    The API uses Redis for caching frequently accessed data. Cache behavior is documented
    per-endpoint. General principles:

    ### Cache Patterns
    | Pattern | TTL | Use Case |
    |---------|-----|----------|
    | `mission:active:{id}` | 5m | Single active mission details |
    | `mission:active:list:{hash}` | 2m | Active missions list with filters |
    | `mission:stats:{scope}:{range}` | 5m | Aggregated statistics |
    | `user:{id}:active_missions` | 2m | User's active missions |

    ### Cache Invalidation
    - Write operations (POST/PUT/DELETE) invalidate relevant cache entries
    - Status changes trigger broadcast invalidation via Redis pub/sub
    - Clients should respect `cached_at` timestamps for cache-aware UX

    ### Endpoints NOT Cached
    - Audit trails (dispatch attempts, inventory logs)
    - Real-time data (tracking, chat)
    - Write operations
    - Paginated history queries

  version: 1.0.0
  contact:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token

  # ============================================================
  # ENUMS - Matching database schema v4.0 exactly
  # ============================================================
  schemas:
...
    UserRole:
      type: string
      enum: [VICTIM, VOLUNTEER, SPONSOR, STAFF, ADMIN]
      description: User role in the system

    # ---------- ENUM: hub_status ----------
...
    MissionType:
      type: string
      enum: [RESCUE, DELIVERY]
      description: Type of mission

    # ---------- ENUM: mission_status ----------
    MissionStatus:
      type: string
      enum: [PENDING, DISPATCHING, ASSIGNED, PICKING_UP, PICKED_UP, IN_TRANSIT, COMPLETED, CANCELLED]
      description: Mission status

    # ---------- ENUM: dispatch_response ----------
...
    VehicleType:
      type: string
      enum: [MOTORBIKE, CAR, TRUCK, BICYCLE, WALK]
      description: Volunteer vehicle type

    # ============================================================
    # COMMON SCHEMAS
    # ============================================================
...
    Error:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input data
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
        phone_number:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        is_verified:
          type: boolean
        is_active:
          type: boolean
        avatar_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Mission:
      type: object
      description: Mission entity matching missions table
      properties:
        id:
          type: string
          format: uuid
        mission_type:
          $ref: '#/components/schemas/MissionType'
        sos_request_id:
          type: string
          format: uuid
          nullable: true
          description: For RESCUE missions
        aid_request_id:
          type: string
          format: uuid
          nullable: true
          description: For DELIVERY missions
        volunteer_id:
          type: string
          format: uuid
          nullable: true
        hub_id:
          type: string
          format: uuid
          nullable: true
          description: For DELIVERY missions
        status:
          $ref: '#/components/schemas/MissionStatus'
        qr_code_token:
          type: string
          nullable: true
        priority_score:
          type: number
          format: double
          nullable: true
        version:
          type: integer
          description: Optimistic locking version
        accepted_at:
          type: string
          format: date-time
          nullable: true
        picked_up_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        cancelled_at:
          type: string
          format: date-time
          nullable: true
        cancellation_reason:
          type: string
          nullable: true
        confirmation_image_url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    MissionCompleteResponse:
      type: object
      description: |
        Response for mission completion.
        Cache: Removes mission:{id} from active cache.
        Triggers:
          - Supabase Realtime status update
          - FCM notification to victim
          - Update volunteer stats in PostgreSQL
      properties:
        success:
          type: boolean
        message:
          type: string
          example: "Mission completed successfully"
        data:
          allOf:
            - $ref: '#/components/schemas/Mission'
            - type: object
              properties:
                completed_at:
                  type: string
                  format: date-time
                confirmation_image_url:
                  type: string
                  description: URL of uploaded confirmation image
                volunteer_stats:
                  type: object
                  properties:
                    total_missions_completed:
                      type: integer
                    points_earned:
                      type: integer
    MissionCancelResponse:
      type: object
      description: |
        Response for mission cancellation.
        Cache: Removes mission:{id} from active cache.
        Triggers:
          - Supabase Realtime status update
          - FCM notification to relevant parties
          - Re-dispatch if victim still needs help (for RESCUE missions)
      properties:
        success:
          type: boolean
        message:
          type: string
          example: "Mission cancelled"
        data:
          allOf:
            - $ref: '#/components/schemas/Mission'
            - type: object
              properties:
                cancelled_at:
                  type: string
                  format: date-time
                cancellation_reason:
                  type: string
                will_redispatch:
                  type: boolean
                  description: Whether system will attempt to find another volunteer
...
    VolunteerProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        is_online:
          type: boolean
        current_lat:
          type: number
          nullable: true
        current_lng:
          type: number
          nullable: true
        vehicle_type:
          $ref: '#/components/schemas/VehicleType'
        total_tasks_completed:
          type: integer
        avg_rating:
          type: number
          format: double
        avg_response_seconds:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    VolunteerProfileResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          allOf:
            - $ref: '#/components/schemas/VolunteerProfile'
            - type: object
              properties:
                user:
                  $ref: '#/components/schemas/UserProfile'
    UpdateVolunteerProfileRequest:
      type: object
      properties:
        vehicle_type:
          $ref: '#/components/schemas/VehicleType'
    VolunteerStatisticsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            total_tasks_completed:
              type: integer
            rescue_missions:
              type: integer
            delivery_missions:
              type: integer
            avg_rating:
              type: number
            total_ratings:
              type: integer
            avg_response_seconds:
              type: integer
            people_helped:
              type: integer
...
paths:
...
  /missions/{id}/complete:
    post:
      tags: [Missions]
      summary: Complete mission with confirmation
      description: |
        Volunteer marks the mission as completed with a confirmation image.
        Required for both RESCUE and DELIVERY missions.

        **Cache Strategy:**
        - Removes: `mission:{id}` from active cache in Redis
        - Mission moves to completed state (no longer cached)

        **Side Effects:**
        - Uploads confirmation image to Cloud Storage
        - Broadcasts status update via Supabase Realtime
        - Sends FCM notification to victim
        - Updates volunteer stats (total_tasks_completed, points)
        - Updates SOS/Aid request status to RESOLVED
      operationId: completeMission
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [confirmation_image]
              properties:
                confirmation_image:
                  type: string
                  format: binary
                  description: Photo proof of delivery/rescue completion
                notes:
                  type: string
                  description: Optional notes about the completion
      responses:
        '200':
          description: Mission completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MissionCompleteResponse'
        '400':
          description: Invalid mission state or missing confirmation image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Mission not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /missions/{id}/cancel:
    post:
      tags: [Missions]
      summary: Cancel mission
      description: |
        Cancel an active mission. Can be initiated by volunteer or staff.

        **Cache Strategy:**
        - Removes: `mission:{id}` from active cache in Redis
        - Mission moves to cancelled state (no longer cached)

        **Side Effects:**
        - Broadcasts status update via Supabase Realtime
        - Sends FCM notification to victim
        - For RESCUE missions: triggers re-dispatch if sos_request still active
        - Returns reserved inventory to hub (for DELIVERY missions)
      operationId: cancelMission
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cancellation_reason]
              properties:
                cancellation_reason:
                  type: string
                  description: Reason for cancellation
                  example: "Vehicle breakdown"
      responses:
        '200':
          description: Mission cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MissionCancelResponse'
        '400':
          description: Invalid mission state (already completed or cancelled)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Mission not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /volunteers/profile:
    get:
      tags: [Volunteers]
      summary: Get volunteer profile
      operationId: getVolunteerProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Volunteer profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VolunteerProfileResponse'

    patch:
      tags: [Volunteers]
      summary: Update volunteer profile
      operationId: updateVolunteerProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVolunteerProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VolunteerProfileResponse'
  /volunteers/status:
    post:
      tags: [Volunteers]
      summary: Toggle online status
      operationId: toggleVolunteerStatus
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [is_online]
              properties:
                is_online:
                  type: boolean
                current_lat:
                  type: number
                  format: double
                current_lng:
                  type: number
                  format: double
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      is_online:
                        type: boolean
                      message:
                        type: string
...
  /volunteers/statistics:
    get:
      tags: [Volunteers]
      summary: Get volunteer statistics
      operationId: getVolunteerStatistics
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VolunteerStatisticsResponse'
...